requirements.txt
groq>=0.13.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
streamlit>=1.33.0
requests>=2.31.0
pandas>=2.2.0
python-dotenv>=1.0.0
